test: move format docstring classes to separate files
